change type data on pengiriman

+++ database/migrations/2023_01_28_044039_create_pengiriman_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('pengiriman', function (Blueprint $table) {
             $table->id();
             $table->string('nama');
-            $table->double('biaya');
+            $table->bigInteger('biaya');
             $table->timestamps();
         });
     }
